+++ OpenMarket/OpenMarket/Protocol/ProductsDataSourceDelegate.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol ProductsDataSourceDelegate: NSObject {
+    func stopActivityIndicator()
+    func showAlert(title: String?, message: String?)
+    func reloadData()
+}
